+++ v8-fibjs/exlib/include/utils_x86.h
@@ -37,9 +37,9 @@ inline T *CompareAndSwap(T *volatile *ptr, T *old_value, T *new_value)
     return prev;
 }
 
-inline int32_t CompareAndSwap(volatile int32_t *ptr, int32_t old_value, int32_t new_value)
+inline intptr_t CompareAndSwap(volatile intptr_t *ptr, intptr_t old_value, intptr_t new_value)
 {
-    int32_t prev;
+    intptr_t prev;
     __asm__ volatile(
         "lock; cmpxchgl %1,%2"
         : "=a" (prev)
@@ -48,9 +48,9 @@ inline int32_t CompareAndSwap(volatile int32_t *ptr, int32_t old_value, int32_t 
     return prev;
 }
 
-inline int32_t atom_add(volatile int32_t *dest, int32_t incr)
+inline intptr_t atom_add(volatile intptr_t *dest, intptr_t incr)
 {
-    int32_t ret;
+    intptr_t ret;
     __asm__ volatile(
         "lock; xaddl %0,(%1)"
         : "=r"(ret)
@@ -59,9 +59,9 @@ inline int32_t atom_add(volatile int32_t *dest, int32_t incr)
     return ret + incr;
 }
 
-inline int32_t atom_xchg(volatile int32_t *ptr, int32_t new_value)
+inline intptr_t atom_xchg(volatile intptr_t *ptr, intptr_t new_value)
 {
-    int32_t prev;
+    intptr_t prev;
 
     __asm__ volatile(
         "lock xchgl %2,(%1)"
@@ -86,12 +86,12 @@ inline T *atom_xchg(T *volatile *ptr, T *new_value)
     return prev;
 }
 
-inline int32_t atom_inc(volatile int32_t *dest)
+inline intptr_t atom_inc(volatile intptr_t *dest)
 {
     return atom_add(dest, 1);
 }
 
-inline int32_t atom_dec(volatile int32_t *dest)
+inline intptr_t atom_dec(volatile intptr_t *dest)
 {
     return atom_add(dest, -1);
 }
